+++ src/app_file.c
@@ -1,28 +1,28 @@
 #include "app_file.h"
 #include "utility.h"
 
-BOOL_INT32 fileWrite_Word(FILE *dev, uint16_t addr, uint16_t value)
+BOOL_INT32 fileWrite_Word(FILE *dev, uint32_t addr, uint16_t value)
 {
 	// TODO
 	debugPrintf("Write words to file!");
 	return BOOL_TRUE;
 }
 
-BOOL_INT32 fileRead_Word(FILE *dev, uint16_t addr, uint16_t *value)
+BOOL_INT32 fileRead_Word(FILE *dev, uint32_t addr, uint16_t *value)
 {
 	// TODO
 	debugPrintf("Read words from file!");
 	return BOOL_TRUE;
 }
 
-BOOL_INT32 fileWrite_Byte(FILE *dev, uint16_t addr, uint8_t value)
+BOOL_INT32 fileWrite_Byte(FILE *dev, uint32_t addr, uint8_t value)
 {
 	// TODO
 	debugPrintf("Write bytes to file!");
 	return BOOL_TRUE;
 }
 
-BOOL_INT32 fileRead_Byte(FILE *dev, uint16_t addr, uint8_t *value)
+BOOL_INT32 fileRead_Byte(FILE *dev, uint32_t addr, uint8_t *value)
 {
 	// TODO
 	debugPrintf("Read bytes from file!");
